+++ src/main/java/com/yongfill/server/domain/file/service/FileUploadService.java
@@ -1,21 +1,25 @@
-package com.yongfill.server.domain.file;
+package com.yongfill.server.domain.file.service;
 
 import com.amazonaws.SdkClientException;
 import com.amazonaws.services.s3.AmazonS3;
+import com.amazonaws.services.s3.model.AmazonS3Exception;
+import com.amazonaws.services.s3.model.CannedAccessControlList;
 import com.amazonaws.services.s3.model.ObjectMetadata;
+import com.amazonaws.services.s3.model.PutObjectRequest;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.util.UUID;
 
 @RequiredArgsConstructor
 @Service
-public class S3FileUploadService {
+public class FileUploadService {
 
     private final AmazonS3 s3Client;
 
@@ -24,6 +28,8 @@ public class S3FileUploadService {
 
     private String defaultUrl = "https://s3.amazonaws.com/";
 
+
+    @Transactional
     public String uploadFile(MultipartFile file) throws IOException {
         if (file.isEmpty()) {
             throw new FileNotFoundException("업로드된 파일이 비어 있습니다.");
@@ -33,7 +39,8 @@ public class S3FileUploadService {
         try {
             s3Client.putObject(bucketName, fileName, file.getInputStream(), getObjectMetadata(file));
             return defaultUrl + fileName;
-        } catch (SdkClientException e) {
+        }
+        catch (SdkClientException e) {
             throw new IOException("Error uploading file to S3", e);
         }
     }
@@ -48,5 +55,21 @@ public class S3FileUploadService {
     private String generateFileName(MultipartFile file) {
         return UUID.randomUUID().toString() + "-" + file.getOriginalFilename();
     }
+
+    private String putS3(File uploadFile, String fileName) {
+        s3Client.putObject(new PutObjectRequest(bucketName, fileName, uploadFile)
+                .withCannedAcl(CannedAccessControlList.PublicRead));
+        return getS3(bucketName, fileName);
+    }
+
+    private String getS3(String bucket, String fileName) {
+        return s3Client.getUrl(bucket, fileName).toString();
+    }
+
+    private void removeFile(File file) {
+        file.delete();
+    }
+
+
 
 }
